refactor: API 도메인 모듈 분리
- user, product 도메인으로 분리
