+++ fe/src/api/product/index.ts
@@ -1,0 +1,57 @@
+import { fetcher } from "@api/fetcher";
+import { AddressList, CategoryInfo, ProductDetailInfo } from "@api/type";
+import { PRODUCT_API_PATH } from "./constants";
+
+export const getAddresses = async (page: number = 0, size: number = 10) => {
+  const { data } = await fetcher.get<AddressList>(
+    PRODUCT_API_PATH.addresses(page, size)
+  );
+  return data;
+};
+
+export const getCategories = async () => {
+  const { data } = await fetcher.get<CategoryInfo[]>(
+    PRODUCT_API_PATH.categories
+  );
+  return data;
+};
+
+export const getProductDetail = async (productId: number) => {
+  return await fetcher.get<ProductDetailInfo>(
+    PRODUCT_API_PATH.productDetail(productId)
+  );
+};
+
+export const postProduct = async (productInfo: FormData) => {
+  const config = {
+    headers: {
+      "content-type": "multipart/form-data",
+    },
+  };
+
+  return await fetcher.post<{ productId: number }>(
+    PRODUCT_API_PATH.newProduct,
+    productInfo,
+    config
+  );
+};
+
+export const patchProduct = async ({
+  productId,
+  productInfo,
+}: {
+  productId: number;
+  productInfo: FormData;
+}) => {
+  const config = {
+    headers: {
+      "content-type": "multipart/form-data",
+    },
+  };
+
+  return await fetcher.patch(
+    PRODUCT_API_PATH.editProduct(productId),
+    productInfo,
+    config
+  );
+};
