+++ fe/src/api/product/queries.ts
@@ -1,11 +1,5 @@
-import { useInfiniteQuery, useQueries, useQuery } from "@tanstack/react-query";
-import {
-  getAddresses,
-  getCategories,
-  getMember,
-  getMemberAddress,
-  getProductDetail,
-} from ".";
+import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
+import { getAddresses, getCategories, getProductDetail } from ".";
 
 export const useAddressesInfiniteQuery = (option?: { size: number }) => {
   return useInfiniteQuery({
@@ -13,27 +7,6 @@ export const useAddressesInfiniteQuery = (option?: { size: number }) => {
     queryFn: ({ pageParam }) => getAddresses(pageParam, option?.size),
     getNextPageParam: (lastPage, allPages) =>
       lastPage.hasNext ? allPages.length : undefined,
-  });
-};
-
-export const useUserInfoQuery = (
-  { enabled }: { enabled: boolean } = { enabled: true }
-) => {
-  return useQueries({
-    queries: [
-      {
-        queryKey: ["getMember"],
-        queryFn: getMember,
-        enabled,
-        staleTime: Infinity,
-      },
-      {
-        queryKey: ["getMemberAddresses"],
-        queryFn: getMemberAddress,
-        enabled,
-        staleTime: Infinity,
-      },
-    ],
   });
 };
