+++ fe/src/components/Modal/AddressModal/AddressModal.tsx
@@ -1,8 +1,8 @@
+import { putUserAddress } from "@api/user";
 import { useToast } from "@hooks/useToast";
 import { useMutation } from "@tanstack/react-query";
 import { isSameItems } from "@utils/index";
-import { putUserAddress } from "api";
-import { useState } from "react";
+import { useRef, useState } from "react";
 import { useAddressList } from "store";
 import Modal from "../Modal";
 import AddressIndicatorList from "./Content/AddressIndicatorList";
@@ -18,8 +18,8 @@ export default function AddressModal({
   const [addresses, setAddresses] = useAddressList();
   const userAddressIDs = addresses.map(({ id }) => id);
 
-  const [prevAddresses] = useState(addresses);
-  const prevAddressIDs = prevAddresses.map(({ id }) => id);
+  const prevAddresses = useRef(addresses);
+  const prevAddressIDs = prevAddresses.current.map(({ id }) => id);
 
   const [isSearchingAddress, setIsSearchingAddress] = useState<boolean>(false);
   const openAddressSearch = () => setIsSearchingAddress(true);
@@ -38,7 +38,7 @@ export default function AddressModal({
         title: "동네 설정 실패",
         message: "동네 설정에 실패했습니다. 잠시 후 다시 시도해주세요.",
       }),
-        setAddresses(prevAddresses);
+        setAddresses(prevAddresses.current);
     },
   });
